+++ src/spellmanager.cpp
@@ -143,7 +143,8 @@ void SpellManager::invoke(const int spellId) const
         >= static_cast<signed>(spell->getBaseLvl()) &&
         PlayerInfo::getSkillLevel(static_cast<int>(
         spell->getSchool())) >= static_cast<signed>(spell->getSchoolLvl())
-        && PlayerInfo::getAttribute(Attributes::MP) >= spell->getMana())
+        && PlayerInfo::getAttribute(Attributes::MP)
+        >= static_cast<int>(spell->getMana()))
         )
 #endif
     {
